refactor(architecture): infraパッケージをinfra/配下に移動
- embedded → infra/embedded
- errorhandler → infra/errorhandler
- loader → infra/loader
- persistence → infra/persistence
- startup → infra/startup

requirements.mdの目標構造に合わせてディレクトリ構造を統一。
全インポートパスを更新し、ビルド・テスト通過確認済み。
