+++ internal/app/message_handlers_test.go
@@ -4,7 +4,7 @@ package app
 import (
 	"testing"
 
-	"hirorocky/type-battle/internal/embedded"
+	"hirorocky/type-battle/internal/infra/embedded"
 	"hirorocky/type-battle/internal/tui/screens"
 
 	tea "github.com/charmbracelet/bubbletea"
