+++ internal/infra/embedded/data/cores.json
