+++ internal/infra/embedded/data/enemies.json
