+++ internal/infra/embedded/data/modules.json
