+++ internal/integration_test/refactoring_verification_test.go
@@ -11,8 +11,8 @@ import (
 	gamestate "hirorocky/type-battle/internal/app/game_state"
 	"hirorocky/type-battle/internal/config"
 	"hirorocky/type-battle/internal/domain"
-	"hirorocky/type-battle/internal/persistence"
-	"hirorocky/type-battle/internal/startup"
+	"hirorocky/type-battle/internal/infra/persistence"
+	"hirorocky/type-battle/internal/infra/startup"
 	"hirorocky/type-battle/internal/tui/screens"
 	"hirorocky/type-battle/internal/usecase/battle"
 	"hirorocky/type-battle/internal/usecase/reward"
