+++ internal/usecase/reward/reward.go
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	"hirorocky/type-battle/internal/domain"
-	"hirorocky/type-battle/internal/loader"
+	"hirorocky/type-battle/internal/infra/loader"
 	"hirorocky/type-battle/internal/usecase/inventory"
 
 	"github.com/google/uuid"
