+++ internal/usecase/reward/reward_test.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	"hirorocky/type-battle/internal/domain"
-	"hirorocky/type-battle/internal/loader"
+	"hirorocky/type-battle/internal/infra/loader"
 	"hirorocky/type-battle/internal/usecase/inventory"
 )
